Moved CapitalizePhrase and EncryptPhrase to common_utils file
